backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
replicate
python-dotenv
openai
sqlalchemy
alembic
aiosqlite
PyJWT
greenlet
